[Feat] 로딩 뷰 구현, 모집글 방장 기능 수정

+++ TrackUs-iOS/TrackUs-iOS/Presentation/Mate/Controllers/CourseDetailVC.swift
@@ -151,8 +151,6 @@ class CourseDetailVC: UIViewController {
     
     private lazy var courseExitButton: UIButton = { // 트랙 나가기 버튼
         let button = UIButton(type: .system)
-        button.backgroundColor = .mainBlue
-        button.setTitle("트랙 나가기", for: .normal)
         button.titleLabel?.textAlignment = .center
         button.titleLabel?.font = UIFont.boldSystemFont(ofSize: 16)
         button.setTitleColor(.white, for: .normal)
@@ -196,6 +194,8 @@ class CourseDetailVC: UIViewController {
         return button
     }()
     
+    var ownerUid: String = ""
+    
     let buttonStack = UIStackView()
     
     lazy var preMapViewTapGesture = UITapGestureRecognizer(target: self, action: #selector(goCourseDetail(_:)))
@@ -211,6 +211,8 @@ class CourseDetailVC: UIViewController {
         collectionView.dataSource = self
         configureUI()
         runningStyleColor()
+        
+        print("DEBUG: OwnerUid = \(ownerUid)")
     }
     
     // MARK: - Selectors
@@ -231,25 +233,33 @@ class CourseDetailVC: UIViewController {
     }
     
     @objc func courseExitButtonTapped() {
-        PostService().exitPost(postUid: postUid, userUid: uid, members: members) { updateMembers in
-            self.members = updateMembers
+        
+        if ownerUid == uid {
+            PostService().deletePost(postUid: self.postUid, imageUrl: self.imageUrl) {
+                self.navigationController?.popToRootViewController(animated: true)
+            }
+        } else {
+            PostService().exitPost(postUid: postUid, userUid: uid, members: members) { updateMembers in
+                self.members = updateMembers
+            }
         }
     }
     
     @objc func goChatRoomButtonTapped() {
+        
+        if members.contains(uid) {
+            // 채팅방에 참여 했을 시 -> 해당 모집글 톡방
+            
+        } else {
+            // 채팅방에 참여 하지 않았을 경우 -> 방장 개인톡
+            
+        }
         
     }
     
     @objc func menuButtonTapped() {
         let editAction = UIAlertAction(title: "모집글 수정", style: .default) { action in
             
-        }
-        
-        let deleteAction = UIAlertAction(title: "모집글 삭제", style: .destructive) { action in
-            
-            PostService().deletePost(postUid: self.postUid, imageUrl: self.imageUrl) {
-                self.navigationController?.popToRootViewController(animated: true)
-            }
         }
         
         let reportAction = UIAlertAction(title: "모집글 신고", style: .destructive) { action in
@@ -260,10 +270,9 @@ class CourseDetailVC: UIViewController {
         
         let alert = UIAlertController(title: nil, message: nil, preferredStyle: .actionSheet)
         
-        if members[0] == uid {
-            // 참여인원의 0번째(작성자)라면
+        if ownerUid == uid {
+            // 작성자인 경우
             alert.addAction(editAction)
-            alert.addAction(deleteAction)
         } else {
             alert.addAction(reportAction)
         }
@@ -378,7 +387,6 @@ class CourseDetailVC: UIViewController {
         buttonStack.distribution = .fill // dldl
         
         // 해당 유저가 참여했는지 안했는지
-        
         if members.contains(uid) {
             buttonStack.addArrangedSubview(courseExitButton)
             buttonStack.widthAnchor.constraint(equalToConstant: 335).isActive = true
@@ -387,10 +395,22 @@ class CourseDetailVC: UIViewController {
             buttonStack.addArrangedSubview(goChatRoomButton)
             goChatRoomButton.widthAnchor.constraint(equalToConstant: 56).isActive = true
             goChatRoomButton.heightAnchor.constraint(equalToConstant: 56).isActive = true
+            
+            if ownerUid == uid {
+                courseExitButton.backgroundColor = .caution
+                courseExitButton.setTitle("글 삭제하기", for: .normal)
+            } else {
+                courseExitButton.backgroundColor = .caution
+                courseExitButton.setTitle("트랙 나가기", for: .normal)
+            }
         } else {
             buttonStack.addArrangedSubview(courseEnterButton)
             buttonStack.widthAnchor.constraint(equalToConstant: 335).isActive = true
             buttonStack.heightAnchor.constraint(equalToConstant: 56).isActive = true
+            
+            buttonStack.addArrangedSubview(goChatRoomButton)
+            goChatRoomButton.widthAnchor.constraint(equalToConstant: 56).isActive = true
+            goChatRoomButton.heightAnchor.constraint(equalToConstant: 56).isActive = true
             
             if members.count >= memberLimit {
                 courseEnterButton.backgroundColor = .systemGray
@@ -453,10 +473,22 @@ class CourseDetailVC: UIViewController {
             buttonStack.addArrangedSubview(goChatRoomButton)
             goChatRoomButton.widthAnchor.constraint(equalToConstant: 56).isActive = true
             goChatRoomButton.heightAnchor.constraint(equalToConstant: 56).isActive = true
+            
+            if ownerUid == uid {
+                courseExitButton.backgroundColor = .caution
+                courseExitButton.setTitle("글 삭제하기", for: .normal)
+            } else {
+                courseExitButton.backgroundColor = .caution
+                courseExitButton.setTitle("트랙 나가기", for: .normal)
+            }
         } else {
             buttonStack.addArrangedSubview(courseEnterButton)
             buttonStack.widthAnchor.constraint(equalToConstant: 335).isActive = true
             buttonStack.heightAnchor.constraint(equalToConstant: 56).isActive = true
+            
+            buttonStack.addArrangedSubview(goChatRoomButton)
+            goChatRoomButton.widthAnchor.constraint(equalToConstant: 56).isActive = true
+            goChatRoomButton.heightAnchor.constraint(equalToConstant: 56).isActive = true
             
             if members.count >= memberLimit {
                 courseEnterButton.backgroundColor = .systemGray
@@ -483,7 +515,7 @@ extension CourseDetailVC: UICollectionViewDelegate, UICollectionViewDataSource, 
         }
         
         let memberUid = members[indexPath.item]
-        let isOwner = indexPath.item == 0
+        let isOwner = ownerUid == memberUid
         cell.configure(uid: memberUid, isOwner: isOwner)
         return cell
     }
